Refresh widget after canceling "Merge When Pipeline Succeeds"
Previously when a user clicked on "Cancel automatic merge", the merge
request widget store would fail with `currentUser is undefined` because
the `cancel_auto_merge` endpoint did not return user information.

Even after handling that problem, the merge request widget did not
refresh until waiting for a while. We fix this by forcing a refresh
after the cancellation is processed.

Closes https://gitlab.com/gitlab-org/gitlab/issues/197487