Stringify sidekiq job args in logs
If these logs are sent to Elasticsearch, it will not be able to process
nested object fields, as this causes a type mismatch with scalar
elements in the same array across log lines.

This is a second attempt, as the first (reverted) one modified the
actual job object that was used by sidekiq.