Make dashboard embedded check more explicit
When a metrics dashboard is requested, the key :embedded
in the query params is used to determine whether a request
is for an embed or a full dashboard. Instead of relying
on just the presence of the param, this swaps the check
over to whether the param's value represents true/false.