Fix inconsistent inflection of Ldap module
Changes modules to all use Ldap instead of LDAP

It was actually easier to change Ldap to LDAP but controllers
don't work because the router uses ActiveSupport inflections. We
could change the ActiveSupport inflection for this but it would
affect multi-word classes like LdapSyncWorker and having to change
all these would be a larger change.