Audit change model that is different than entity
This extends `for_changes` method to accept reference of target object
being audited. Previously the `entity` is used to record both scope and
target change object. While that is true in some context (i.e. the
target model happens to be a User/Project/Group), it does not work for
`ApprovalProjectRule` (i.e. the target model is `ApprovalProjectRule`
and the entity is the associated Project)