Ensure RepositoryLinkFilter handles Gitaly failures gracefully
We do this by setting all paths to `:blob` type if we get either
`GRPC::Unavailable` or `GRPC::DeadlineExceeded` when calling
`GitalyClient`. This is acceptable because all files existing in a
repository can be viewed in "blob" mode, although if Gitaly is down we
won't be able to fetch the object anyways.