Remove unnecessary Jira template test
When a template is applied to a project
it copies all attributes to a new service
and sets template to `false` and project_id
to the ID of the project. After the service was
copied it is not different from any other
project service and behaves the same.

A service should not be a template and
attached to a project at the same time

To simulate a template applied to a project
we would need to use Projects::CreateService.
But this test would not belong to
services_controller_spec.