Fix flaky user_manages_subscription_spec.rb
It looks like the feature spec was attempting to click on the
subscription button before all the Ajax calls for the page had
loaded. As a result, the click did nothing, and the test failed.

To fix this, we should wait for outstanding requests to be ready before
attempting to click.

Closes https://gitlab.com/gitlab-org/gitlab/issues/207835