Fix k8s deployment namespace resolution
The namespace shown on the job show page was sometimes incorrect for
deployments to non-managed clusters when
environment:kubernetes:namespace was specified in the CI YAML. This was
due to the last successful deployment being used to resolve the CI YAML
namespace, instead of the CI YAML for the deployment being constructed.

Note that the correct namespace was still being used for the actual
deployment.