BulkIndexer change _routing -> routing to support ES 7
In ES 6.x onwards _routing is deprecated in the bulk API but in 7.x it
was removed completely. While this is sadly not clearly documented it's
referenced in a few places like
https://discuss.elastic.co/t/deprecated-field-routing-used-expected-routing-instead/115951
and we also learnt this when fixing the Go indexer at
https://gitlab.com/gitlab-org/gitlab-elasticsearch-indexer/-/merge_requests/34

I have tested this version works for 6.5.1 , 6.8.1 and 7.5.1 .
Additionally since this is the same way that our
Go gitlab-elasticsearch-indexer is behaving this should be a safe
change.

See https://gitlab.com/gitlab-org/gitlab/-/issues/208717#note_300201720