Introduce workers to sync Seat Link information
The SyncSeatLinkWorker gathers the max historical user count from the
start of the license to the previous day, then executes the
SyncSeatLinkRequestWorker. The SyncSeatLinkWorker is a Sidekiq
CronWorker that can be retried up to 12 times (~17 hours). It can be
retried past the original job creation day to report the correct day.

The SyncSeatLinkRequestWorker is responsible for making request to the
Subscription Portal with seat info. Because all the information
necessary to make the request is pulled from the job information, this
job can be retried beyond the original job creation day.

Using UTC time to determine when to sync seat information and what
time range to use for gathering the max user count.