Fix SCIM API Application Context
The SCIM API needs
to override the Application Context and set the user to nil
since the SCIM API doesn't operate in the context of a user.
Rather, the SCIM API uses a static token specific to the group
enabling SCIM.

Without overriding the context saving a user will hang when
attempting to resolve the current_user set in the Application
Context in the main API (lib/api/api.rb).