Geo Replication - Paramaterized API
This MR is an effort at MVC.

This MR splits off from:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/26410

The overall focus of this and following related
changes is to replace Geo_dDesigns with
Geo_Replicable.

This MR adds the prop "replicable_type".
This prop is used in the geo_replication api
to denote which type of replicable we are using.

It also removes an un-used prop:
"design_management_link" that had no
usage in the components.

Following MRs:
Public Store Functions and State
Vue Components and Filenames