Make milestones issue counters independent of user
We want to expose milestone issue counts to the API for
https://gitlab.com/gitlab-org/gitlab/issues/31289.

The current milestones page is also very slow -
it has N+1 for the same counters.

We want to cache these counters, but they depend on user,
which makes cache useless.

So this commit removes dependency on `user` variable for
milestone issue counters