Extract AdvancedStageWorker to be reused
AdvanceStageWorker is a worker that initally was
used by github importer.

Extracted its functionality so that it can be used by
any multi stage importer to wait for a number of jobs
to complete, without blocking a thread. Once all jobs
have been completed this worker will advance the import
process to the next stage.