Ability to distinguish if user is Gitlab employee
Adds an instance method on the user model to determine
whether the user is a GitLab employee or not.

It also exposes the is_gitlab_employee attribute in the
UserBasic entity used in discussion api to use
this detail to show a GitLab badge to discussion entities
of the company's employees.