Use tmp/gitlab_workhorse_test_secret for tests
Originally we used tmp/tests/gitlab_workhorse_secret, but this is in a
.gitignore'd directory and this fails to transfer to gitlab-foss
correctly.

We don't seem to be able to manually fix it up in gitlab-foss, so
remove the missing .gitkeep file in this project instead, and use
tmp/gitlab_workhorse_test_secret instead.

The problem exists because we need the secret to exist very early on
in initialization, and I'm not particularly comfortable with rails
creating directories willy-nilly.