Re-design Geo Replication Popover
Currently it is a little confusing from the
Geo Node Status screen, to see in detail
what is going on with your syncs.

To comabt that, we have been working on
ways to make the UI a little more intuitive.

This MR focuses on the popover that
shows up when you hover over a sync status.

Now it will break where different items are
in the sync, as well as provide a "More Information"
link when available.