Backfill LfsObjectsProject records of forks
To fix the behavior of existing forks, LfsObjectsProject records
need to be backfilled. This is so we can remove the need to depend
on the source when looking for LFS objects from forks.

This is a variation of https://gitlab.com/gitlab-org/gitlab/-/merge_requests/25343
that was reverted. The difference is that we are now enqueueing
background migration jobs based on lfs_objects_project records.