Allow downloading of daily coverage
This is the backend work for the download feature.
The contoller exposes options for the frontend to use for rendering
the download button and later on the actual code coverag graph.

Also, adds a single batch csv builder into its own class.

This simplifies things and makes intent clearer when using a
separate class for single batch CSV builder.