Run QA frontend jobs when dependencies are updated
The QA jobs purpose is to catch if we utilize JavaScript that cannot be
executed on one of the node LTS versions we support for from-source
builds.

Currently we execute the QA jobs on master.

1. This means that they are executed unnecessarily, e.g. on docs only
   branches.
2. The thing which breaks the QA jobs are likely dependency updates, as
   we just test the webpack compilation and dependencies and
   subdependencies are usually requiring newer node versions. So we
   should be executing these checks on branches that update
   dependencies.

Recently we had a red master, because one dependency update required a
newer node version and the pipeline was green [0], [1].

Therefore I am proposing that we run the QA frontend jobs in merge
requests which are changing our package.json or yarn.lock file, as we
want to know if a dependency update would fail our build _before_ a red
master happens.

[0]: https://gitlab.com/gitlab-org/gitlab/-/issues/202187
[1]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/24413