Refactor some Geo Node tests with vue-test-utils
This MR is split off from:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/27033

In the above MR we implement GlPopover.  However,
in doing so there are problems with fully mounting it.

Using shallowMount and stubbing it fixes the problem,
however, the affected tests were written without
vue-test-utils and thus cannot shallowMount.

This MR simply rewrites the impacted tests with
vue-test-utils and shallowMounts the components.