Fix flaky create_snippet_spec.rb
This spec was failing intermittently because the empty state SVG would
be lazily loaded, causing the "New snippet" button to shift
momentarily. Capybara would attempt to click on the coordinates of the
original button, but nothing would happen.

To make this test more robust, wait for the image to load before
attempting to click on the button.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/209892