Remove repository import code from Gitlab::Shell
The import_repository GRPC call is used when importing a project from
a remote source. Historically, this was functionality we'd ask a
subprocess in gitlab-shell to do. Now, Gitaly does it for us instead.

Refactor the code so we no longer have to indirect via Gitlab::Shell
to perform this call. It makes what is happening much clearer, and is
a precursor to removing all references to Gitlab::Shell.