Geo Replication - Rename Store/State
This MR is an effort at MVC.

This MR splits off from:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/26410

The overall focus of this and follwoing related
changes is to replace Geo_Designs with
Geo_Replicable.

This MR renames the "public" store actions.
In doing so the component references
were updated as well.

This MR renames the state properties that
we explicit for designs to replicable.

Following MRs:
Rename Vue Components and Filenames