Store the number of scanned URLs on security_scan
The DAST job artifact will include the URLs that DAST
scanned. To display the number of URLs scanned
on the MR I’ll be storing the URL count on
security_scans. This MR adds the column scanned_resources_count
to security_scans and populates that field once a
pipeline completes.