Fix readable_by? method
- Check if the user is a member of the project
if the issue is confidential
- Add specs to cover moved issues

Add specs for ReferenceRedactorFilter

Add check for project access

- When the issue checked in #readable_by?
is confidential the assignee or author should
be able to access the project to view it.

Refactor #readable_by?

- Refactor Issue specs to cover all cases
in readable_by? conditions

Refactor reference_redactor_filter_spec

-  Change empty lines to apply Four-Phase
test pattern

Add specs in IssuePolicy to test moved issue case