Add context to object storage error message
If object storage is not enabled for a certain type of object, project
exports may see "Object storage is not enabled". It's hard to know what
type of object is not enabled, so add the class name of the uploader to
see.

This came out of an attempt to run the project export in Kubernetes.