Use the legacy Gemfile/Gemfile.lock names in tooling/overcommit
This is because when multiple versions of Bundler are installed (e.g.
1.17.3 and 2.1.4), Bundler will choose the 2.x version based on the
`gems.rb` / `gems.locked` file names but it chooses the 1.17.x version
if the we use the legacy `Gemfile` / `Gemfile.lock` file names.

For now, we'll use the legacy file names to be consistent with the main
`gitlab`'s `Gemfile.lock` and to avoid unexpected changes in the
working tree.

Signed-off-by: Rémy Coutable <remy@rymai.me>