Refactor additional Geo Node tests
This MR splits off from:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/27624

In the above MR, GlPopover is implemented.
There are some problems when fully
mounting it in a spec.

Using shallowMount and stubbing it fixes the problem,
however, the affected tests were written without
vue-test-utils and thus cannot shallowMount.

This MR simply rewrites the impacted tests with
vue-test-utils and shallowMounts the components.