Refactor board.js to Vue single file component
Currently the board.js component is rendered based `text/x-template`
which is created in HAML. We want to utilize a single file component
(SFC), as this makes it testable with jest without generating fixtures.

This commit is quite large, we implement the change behind a feature
flag and have quite a lot of confidence thanks to our integration tests
and pinning tests. Changes done in this commit:

- adding a new SFC and it's Enterprise counterpart
- utilize a feature flag to either use the new component or the old.
- Add jest specs for the new component based on old karma specs
- Utilize GitLab UI's Icon component for all icons
- Add two properties to the new component, so that we can pass group IDs
  and whether a user can edit the board column
- Rename board_component to board_column This reflects much better what
  the component is actually about: It is a column in our issue boards.

Co-authored-by: Lukas Eipert <leipert@gitlab.com>
Co-authored-by: Mark Florian <mflorian@gitlab.com>
Co-authored-by: Paul Slaughter <pslaughter@gitlab.com>