Fix broadcast message rendering
With the broadcast placeholders we're using
`Banzai#render_and_post_process` instead of `Banzai#render_field` like
before. `render_and_post_process` is using `render` instead of
`render_field` internally, which means that we need to pass in the text
of a broadcast message instead of which field should be rendered.

However another difference of this method is that it is not
automatically using the `BroadcastMessagePipeline` for rendering. We
could overcome this by passing `pipeline: :broadcast_message` in the
context but I've instead added another method
`render_field_and_post_process` which calls `render_field` internally
instead to align with the behaviour that we used before.