Fix direct access to deprecated design path
71d924603316124a7ba321232ee9514e52a8cf07 introduced a change to issue
routes, deprecating the old ones, but still supporting them. There was
added a special variable in issues/_discussion.html.haml to
inform the Vue app about which of the two possible routes the Vue app
was mounted on.

The calculation of `deprecated_route` was previously only returning
true if the user was looking at:

- issues/1
- issues/1/designs

But not:

- issues/1/designs/mydesign.png

In Rails, everything after the `designs` segment of the route is
wildcarded as a single param[:vueroute] so it can be handled by Vue.

https://gitlab.com/gitlab-org/gitlab/-/issues/211833