Fix updating the authorized_keys file
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/26913, we
merged a change that breaks updating the authorized_keys file. Sidekiq
converts symbol arguments to #perform into string arguments, and we
didn't take account of that. Now updating the authorized_keys file is
broken :S

This commit ensures we always treat the passed-in action argument as a
string, which fixes the issue.