Geo Nodes - Conditional Removal
A primary node is unable to be removed if
a secondary node exists.  However,
the UI doesn't reflect this.

This MR adds logic to disable the
remove button on a Primary node
in the UI if there are additional
nodes.

A tooltip will pop up on the
disabled button with an explanation
why it is disabled.

This MR also moves up the node actions
to be inline with the Node URL.