Add charts to the trigger-build-docs script
This adds the charts project to the list of the projects
that can use the script to trigger review apps for docs.

The slug that's used for each project was defined by the
CI_PROJECT_NAME variable. Now that we're adding charts to
the mix, we need to find a different way, since the charts
project (gitlab-org/charts/gitlab) has the same CI_PROJECT_NAME
as GitLab (gitlab-org/gitlab).

Thus, we need a different way to differentiate the two projects.