Add proxy config note to usage statistics docs
Previously the usage statistics documentation did not mention proxy
configuration. Proxy mis-configuration is a common cause of usage
statistics failing to be reported, which can cause other features
(like DevOps Score) to be non-functional. This adds a brief note about
proxy configuration to the top of the usage statistics documentation.