Enable placeholders for broadcast notifications
This removes the feature flag for broadcast placeholders and only
enables placeholders for notifications. This feature flag changes
the logic intentionally. During initial development it came
apparent that placeholders only make sense for notifications, but
this was then not reflected in the final MR.

This MR fixes this, removes the feature flag, adds a test for the
new code path and cleans up the specs.