Don't try to deduplicate scheduled jobs
Sometimes we schedule jobs in the future. Now we'd mark them as
duplicates when scheduling them if there was already another job in
the queue with similar arguments.

But since the job could be for a long time in the future, no need to
mark it as a duplicate since it might be intentional to repeat the
work.