Find or initialize instance-level services
This is similar to the service templates feature that we are going to
remove: https://gitlab.com/groups/gitlab-org/-/epics/2430 but we don't
need to create the services in the database at this point.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/25752 implemented
the frontend changes under a feature flag, this implements the backend
changes.

This also disables instance_level_integrations feature flag in the
tests.