Use prefix search instead of ngrams for sha fields
Currently, SHAs are indexed using ngrams from 5 to 40 characters. This
means that each SHA is split into 35 separate terms taking up a lot of
storage. SHAs are quite unique from 4-5 characters on, so a simple
prefix search will be sufficiently fast and as effective as ngrams with
term matching. This change replaces current sha fields analyzers with
prefix search.