Include LDAP UID attr in default attrs for LDAP lookups
In cases where the configured LDAP `uid` attribute is non-standard
(not `uid`, `samaccountname`, `userid`) the `username` value in
the LDAP auth hash will not contain a value. By including the
configured `uid` attribute in the `username` default attributes
array we ensure that the value will be present in the auth hash.