Mark ScheduleAggregationWorker as idempotent
This marks the Namespaces::ScheduleAggregationWorker as
idempotent. It was already idempotent, because it checks the existence
of the Namespace::AggregationSchedule before doing anything and it
uses `safe_find_or_create` to avoid race conditions.

This adds specs proving that it is idempotent.

It also adds the worker to the DROPPABLE_QUEUES so we can use it to
try out deduplicating jobs.