Find or initialize instance-level integration
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/26454 implemented
the basic feature under a feature flag, this implements the backend
changes.

This is similar to the service templates feature that we are going to
remove https://gitlab.com/groups/gitlab-org/-/epics/2430. As for service
templates, we are not using the testing settings functionality because
is couple to project that we don't have in this context.

We will enable the testing settings functionality one-by-one
https://gitlab.com/gitlab-org/gitlab/-/issues/213138.