Pass related_class to the context as a string
Since the Gitlab::ApplicationContext wants a string, this means that
anything else will not be passed to Labkit for logging.

We did this to avoid logging huge objects that could contain sensitive
information.

This also caused the ReactiveCachingWorker to miss it's related
class sometimes: We schedule this in
ReactiveCaching#refresh_reactive_cache! using self.class, this
would then filtered out since Class is not a String.