Add endpoints for the group export UI
We want to be able to export groups from the UI as well as the API, in
the same style as project export. Add the two required endpoints to
achieve this.

Also add the functionality to remove group export files when a new
export is triggered (in line with the regenerate behaviour on project
export) so that we do not have redundant files hanging round.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/211805