Add worker label to prometheus metrics
As part of the work to reduce the amount of queues
Sidekiq has to listen to, here we introduce a worker
label that in the future will provide enough information
to query metrics per worker.

Today it's not strictly required given we
have an one-queue-per-worker approach (so we're
able to know which worker is causing issues through
its queue name).