Add support for DB-less embeds
In order to support metrics embeds for alerts from external
alerting sources, we need a way to store that content,
because the alert does not yet live in the database.

This adds a class which allows post-processing for a
url-provided embed json object. As dashboard content is
already user-provided, this poses no further security concerns.